Answer the third review round: exits, clocks, pids, and file sizes
The snapshot child's entry now returns its exit code instead of calling
Deno.exit mid-run, so the child's clean-up — waiting out its last touch
and aging an orphaned claim — really runs. A run's record drops the
child's process id the moment the child ends, so --kill can never trust
an id that may already be somebody else's while the copy-back finishes.
A claim stamped in the future reads as walked away rather than
unstealable, and a claim with no readable time on a filesystem that
keeps no write times fails loudly instead of crashing bare. The claim
suite is split into focused files under test/scripts/stale-claim/, its
timing assertions wait for the looked-for state instead of sampling
fixed pauses, and the taking path reuses the one guard helper. The run
liveness check is renamed runIsOwned to say what it reads.
